add trackingMode option
- replace activateScrollOnTouchEnd with trackingMode option
- remove support-touch and other unused files
- move window.chrome type to dom-not-standard
- add missing end scroll call for touch event handler
