+++ src/api/options/chart-options-defaults.ts
@@ -1,6 +1,6 @@
 import { isRunningOnClientSide } from '../../helpers/is-running-on-client-side';
 
-import { ChartOptionsInternal } from '../../model/chart-model';
+import { ChartOptionsInternal, TrackingExitMode } from '../../model/chart-model';
 
 import { crosshairOptionsDefaults } from './crosshair-options-defaults';
 import { gridOptionsDefaults } from './grid-options-defaults';
@@ -37,7 +37,6 @@ export const chartOptionsDefaults: ChartOptionsInternal = {
 		pressedMouseMove: true,
 		horzTouchDrag: true,
 		vertTouchDrag: true,
-		activateScrollOnTouchEnd: false,
 	},
 	handleScale: {
 		axisPressedMouseMove: {
@@ -51,5 +50,8 @@ export const chartOptionsDefaults: ChartOptionsInternal = {
 	kineticScroll: {
 		mouse: false,
 		touch: true,
+	},
+	trackingMode: {
+		exitMode: TrackingExitMode.OnNextTap,
 	},
 };
