--- src/gui/last-touch-or-mouse-event.ts
@@ -1,19 +1,0 @@
-import { mobileTouch } from './support-touch';
-
-let lastEventIsTouchValue = mobileTouch;
-
-export function lastEventIsTouch(): boolean {
-	return lastEventIsTouchValue;
-}
-
-export function lastEventIsMouse(): boolean {
-	return lastEventIsTouchValue;
-}
-
-// This setter is called from MouseEventHandler on any touch / mouse event.
-// Currently we use MouseEventHandler on PaneWidget / PriceAxisWidget / TimeAxisWidget (almost the whole chart).
-// There are places where we should know what is type of the last mouse or touch event
-// and where MouseEventHandler is not used, for example ChartWidget / CrosshairPaneView.
-export function setLastEventIsTouch(isTouch: boolean): void {
-	lastEventIsTouchValue = isTouch;
-}
