+++ src/typings/_global-types/index.d.ts
@@ -23,18 +23,3 @@ declare const enum MouseEventButton {
 type Mutable<T> = {
 	-readonly [P in keyof T]: T[P];
 };
-
-/** https://developer.mozilla.org/en-US/docs/Web/API/InputDeviceCapabilities */
-interface InputDeviceCapabilities {
-	firesTouchEvents?: boolean;
-}
-interface UIEvent {
-	/** https://developer.mozilla.org/en-US/docs/Web/API/UIEvent/sourceCapabilities */
-	sourceCapabilities?: InputDeviceCapabilities;
-}
-
-interface Window {
-	chrome?: {
-		runtime: unknown;
-	};
-}
